api/types: move AuthConfig to registry types
Making the api types more focused per API type, and the general
api/types package somewhat smaller.

(cherry picked from commit 818ee962196d668b6d405835c88c1ac6cdab7423)
